+++ src/main/java/com/ga/service/UserServiceImpl.java
@@ -71,7 +71,7 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public User addSong(String username, int songId) {
+	public User addSong(String username, Long songId) {
 		return userDao.addSong(username, songId);
 	}
